fix!: update remove_at in LinkedList
The clear method in hash maps was found to not clear the head node in case
of lists containing more than one nodes. The issue was traced back to the
remove_at method in LinkedList class. The removal of head elements was not
handled. This was solved by adding the remove_head method

Commit includes linting corrections as well